perf: isinstance + is_string_compatible
