+++ lazy_type_hint/data_type_tree/generic_type/set_data_type_tree.py
@@ -1,6 +1,6 @@
 from typing import Any, Hashable, Literal, Sequence, Set, Tuple
 
-from typing_extensions import override
+from typing_extensions import Self, override
 
 from lazy_type_hint.data_type_tree.data_type_tree import DataTypeTree
 from lazy_type_hint.data_type_tree.generic_type.generic_data_type_tree import (
@@ -13,6 +13,8 @@ class SetDataTypeTree(GenericDataTypeTree):
     wraps = (frozenset, set)
     children: Sequence[DataTypeTree]
     operations: SetAndSequenceOperations
+
+    _iterator: int
 
     @override
     def __pre_child_instantiation__(self) -> None:
@@ -37,3 +39,17 @@ class SetDataTypeTree(GenericDataTypeTree):
         for child in self:
             hashes.add(child._get_hash())
         return frozenset(hashes)
+
+    @override
+    def __iter__(self) -> "Self":
+        self._iterator = 0  # Reset the index to zero when starting a new iteration
+        return self
+
+    @override
+    def __next__(self) -> "DataTypeTree":
+        if self._iterator < len(self.children):
+            element = list(self.children)[self._iterator]
+            self._iterator += 1
+            return element
+        else:
+            raise StopIteration
